Fix CVE-2024-27304
SQL injection can occur if an attacker can cause a single query or bind
message to exceed 4 GB in size. An integer overflow in the calculated
message size can cause the one large message to be sent as multiple
messages under the attacker's control.

Update to pgproto3/v2 v2.3.3 which checks for too large messages.